add 'align' method
When a Net::Whois::Object::* object is created in the style
Net::Whois::Object::InetNum->new(
    inetnum => $inetnum,
    org     => $org,
    ...
);
the ->dump output is not aligned by default, which means
that unaligned data is entered into the RIPE db, which people
generally don't find very pretty.

This patches adds an ->align method which pads the values with
whitespace to generate nice visual layout.